mapping/pyproject.toml
[project]
name = "mapping"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "Bilal Faye", email = "bilal.faye@pnnl.gov" }, 
    { name = "Majid alDosari", email = "majid.aldosari@pnnl.gov" }, 
]
dependencies = [
    "rdflib",
    "pyyaml",
    "pyshacl",
    "owlrl",
    "oxrdflib",
    "requests-file",
    "fire",
    "tqdm",
]
readme = "README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pandas",
    "shelved-cache",
     ]

[tool.hatch.metadata]
allow-direct-references = true

[tool.rye.scripts]
mapping = "python -m mapping.speckle"
